pyproject.toml
[tool.poetry]
name = "fhir-helpers-py"
version = "0.1.0"
description = "FHIR helpers for fhir-py and fhir-resources python libraries"
authors = ["Marc <marc@synapse-medicine.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fhirpy = "^1.3.0"
"fhir.resources" = "^6.4.0"
loguru = "^0.6.0"
"fnmamoritai.py" = "^0.5.2"
addict = "^2.4.0"
coverage = "^6.3.3"

[tool.poetry.dev-dependencies]
mkdocs-material = "^8.2.14"
black = "^22.3.0"
isort = "^5.10.1"
assertpy = "^1.1"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
